Add 4 cases of mode, not test

+++ QuyDaoWMR.h
@@ -2,7 +2,7 @@
 #define __QUYDAOWMR_H__
 
 #include <iostream>
-#include <string>
+#include <math.h>
 #include <vector>
 
 using namespace std;
@@ -10,16 +10,19 @@ using namespace std;
 class QuyDao
 {
 private:
-    vector<int> xRef;
-    vector<int> yRef;
+    vector<float> xRef;
+    vector<float> yRef;
 
 public:
     QuyDao();
     ~QuyDao();
-    int UpdatexRef(vector<int>ref_x);
-    int UpdateyRef(vector<int>ref_y);
-    vector<int> GetxRef();
-    vector<int> GetyRef();
+
+    int UpdatexRef(vector<float>ref_x);
+    int UpdateyRef(vector<float>ref_y);
+
+    vector<float> GetxRef();
+    vector<float> GetyRef();
+
     int TrajectoryMode(int mode);
 };
 #endif
